fix: 공지사항 데스크톱 헤더의 너비가 맞지 않는 문제 해결 #90
- 별도의 컴포넌트로 만들어 재사용
